+++ NEWS.md
@@ -1,5 +1,7 @@
 # cli (development version)
 
+* New `keypress()` function to read a single keypress from a terminal.
+
 * New function `pretty_print_code()` to print function objects with syntax
   highlighting at the R console.